Add generic ImageFormat backend
This is designed with the goal of taking care of all the array reading
once the metadata/header info is in place. Ideally this could benefit
other packages moving towards supporting Julia Images and support
easy implementation of unique I/O functions at a common bottleneck,
before a unfiform set of elements is read into an array.

Hoping future additions will include:
- eltype transformations based on named dimensinons: I think I'll wait
to do this until NamedDims.jl is in plae because it looks like a
promising way to make this generic and performant.
- Read in directly to different <:AbstractArray (perhaps unique stuff
for segmentations etc.)