Basic writing works now
Although there aren't formal tests for this, I was able to use the
example4d in tests to read-->write-->read and get the same image. The
only caveat is that the forth column and row of teh sform code is not
currently preserved. It doesn't make a difference in terms of
identifying orientation, but I should probably work this out before
finalizing anything.