Passing Tests Now!
Although this is far from perfect, it is passing tests. Mainly, it
covers accurate reading and access to traits for both NIfTI-1/2 and
HDR/IMG formats. Writing should work but I still need to add thorough
tests for this and I need to figure out some problems with the qform
field in NIfTI first