Begin modernizing this thing
Although this started with an attempt to address corruption of `Int16`
bit types, I had to add in some stuff that I've been using privately for
a while that was driving me crazy. So rather than a quick fix this is a
huge refactor that will take a bit more time.